// src/enums/LogGroup.ts
export enum LogGroup {
  ENEMY = 'ENEMY',
  QUIRK = 'QUIRK',
  PLAYER = 'PLAYER',
  GAME = 'GAME',
  UI = 'UI',
  PATHFINDING = 'PATHFINDING',
  POWERUP = 'POWERUP',
  PROJECTILE = 'POWERUP-[PROJECTILE] ',
  MAP = 'MAP',
  SCENE = 'SCENE',
  INPUT = 'INPUT',
  COLLISION = 'COLLISION',
  ANIMATION = 'ANIMATION',
  SOUND = 'SOUND',
  PRELOAD = 'PRELOAD',
  MENU = 'MENU',
  LOCALIZATION = 'LOCALIZATION',
  RENDERER = 'RENDERER',
  PERFORMANCE = 'PERFORMANCE',
  PALETTE = 'PALETTE',
  BONUS = 'BONUS',
  DEVELOPER = 'DEVELOPER MODE',
  TOUCH = 'TOUCH'
}